+++ mains/main.cpp
@@ -133,7 +133,7 @@ int mag_main(int argc, char** argv) {
         while (1) {
 #ifdef _WIN32
             DWORD count =
-                GetModuleFileNameA(NULL, program_name_storage.buffer(), program_name_storage.cap());
+                GetModuleFileNameA(NULL, program_name_storage.buffer(), (DWORD)program_name_storage.cap());
 #else
             ssize_t count = readlink("/proc/self/exe", program_name_storage.buffer(),
                                      program_name_storage.cap());
